рефакторит модуль с утилитами
