+++ src/view/popup-view.js
@@ -1,11 +1,11 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {humanizeDate} from '../utils.js';
-import {formatRuntime} from '../utils.js';
+import {humanizeDate} from '../utils/task.js';
+import {formatRuntime} from '../utils/task.js';
 
 const createPopUpTemplate = (film, comments) => {
   const {title, totalRating, release, runtime, genres, poster, alternativeTitle, description, ageRating, director, writers, actors} = film;
   const {releaseCountry} = release;
-  const releaseDate = release.date !== null ? humanizeDate(release.date, 'DD MMM YYYY') : '';
+  const releaseDate = humanizeDate(release.date, 'DD MMM YYYY');
 
   const createGenreTemplate = (filmGenres) => {
     let template = '';
@@ -20,7 +20,7 @@ const createPopUpTemplate = (film, comments) => {
 
     for (const filmComment of filmComments) {
       const {author, comment, date, emotion} = filmComment;
-      const commentDate = date !== null ? humanizeDate(date, 'YYYY-MM-DD, h:mm') : '';
+      const commentDate = humanizeDate(date, 'YYYY-MM-DD, h:mm');
 
       template += `
       <li class="film-details__comment">
